backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
python-socketio[asyncio]==5.16.0
aiohttp==3.10.10
networkx==3.4.2
httpx==0.27.2